Fix foreground service crash and JSON duration parsing
- Use startService() instead of startForegroundService() — MediaSessionService
  promotes to foreground automatically when playback starts
- Change duration fields from Long to Double to handle yt-dlp float values
- Add isLenient and coerceInputValues to JSON parser
- Add sessionActivity PendingIntent for notification tap

+++ app/src/main/java/com/maestro/android/data/model/Models.kt
@@ -7,7 +7,7 @@ data class Track(
     val id: String,
     val title: String,
     val artist: String? = null,
-    val duration: Long? = null,
+    val duration: Double? = null,
     val thumbnail: String? = null,
     val url: String = ""
 )
